LaunchDarkly.Server.Store.Redis

Description

The public interface for the LaunchDarkly Haskell Redis integration

Documentation

data RedisStoreConfig

Opaque type used to configure the Redis store integration.

makeRedisStoreConfig :: Connection -> RedisStoreConfig

Create a default config from a given connection pool.

redisConfigSetNamespace :: Text -> RedisStoreConfig -> RedisStoreConfig

Configure the Redis key prefix. All keys are prefixed by default before being inserted into Redis. The default prefix is "launchdarkly".

makeRedisStore :: RedisStoreConfig -> IO PersistentDataStore

Construct a PersistentDataStore that can then be used during SDK configuration.